requirements.txt
streamlit
pandas
python-dotenv
plotly
requests
